# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RoomImage {
  id: number;
  roomId: number;
  imageUrl: string;
  altText?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface AddRoomImageInput {
  imageUrl: string;
  altText?: string | null;
  displayOrder?: number;
}

export type RoomType = (typeof RoomType)[keyof typeof RoomType];

export const RoomType = {
  standard: "standard",
  deluxe: "deluxe",
  suite: "suite",
  penthouse: "penthouse",
} as const;

export interface Room {
  id: number;
  name: string;
  type: RoomType;
  pricePerNight: number;
  capacity: number;
  amenities: string[];
  available: boolean;
  description: string;
  imageUrl?: string | null;
  images: RoomImage[];
  createdAt: string;
  updatedAt: string;
}

export type CreateRoomInputType =
  (typeof CreateRoomInputType)[keyof typeof CreateRoomInputType];

export const CreateRoomInputType = {
  standard: "standard",
  deluxe: "deluxe",
  suite: "suite",
  penthouse: "penthouse",
} as const;

export interface CreateRoomInput {
  name: string;
  type: CreateRoomInputType;
  pricePerNight: number;
  capacity: number;
  amenities?: string[];
  available?: boolean;
  description: string;
  imageUrl?: string | null;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  pending: "pending",
} as const;

export interface Booking {
  id: number;
  roomId: number;
  guestName: string;
  guestEmail: string;
  checkIn: string;
  checkOut: string;
  totalPrice: number;
  status: BookingStatus;
  createdAt: string;
  updatedAt: string;
  room?: Room | null;
}

export interface CreateBookingInput {
  roomId: number;
  guestName: string;
  guestEmail: string;
  checkIn: string;
  checkOut: string;
}

export type DiningItemCategory =
  (typeof DiningItemCategory)[keyof typeof DiningItemCategory];

export const DiningItemCategory = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  drinks: "drinks",
  dessert: "dessert",
} as const;

export interface DiningItem {
  id: number;
  name: string;
  description: string;
  category: DiningItemCategory;
  price: number;
  imageUrl?: string | null;
  available: boolean;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateDiningItemInputCategory =
  (typeof CreateDiningItemInputCategory)[keyof typeof CreateDiningItemInputCategory];

export const CreateDiningItemInputCategory = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  drinks: "drinks",
  dessert: "dessert",
} as const;

export interface CreateDiningItemInput {
  name: string;
  description: string;
  category: CreateDiningItemInputCategory;
  price: number;
  imageUrl?: string | null;
  available?: boolean;
  displayOrder?: number;
}

export interface HotelSettings {
  facebookUrl?: string;
  twitterUrl?: string;
  instagramUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  hotelName?: string;
}

export interface AiChatInput {
  message: string;
}

export interface AiChatResponse {
  response: string;
  tool?: string | null;
  data?: unknown | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}