# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hotel Booking System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Room operations
  - name: roomImages
    description: Room image operations
  - name: bookings
    description: Booking operations
  - name: dining
    description: Dining operations
  - name: settings
    description: Hotel settings
  - name: admin
    description: Admin operations
  - name: ai
    description: AI agent operations
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [admin]
      summary: Create a room (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRoom
      tags: [admin]
      summary: Update a room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomInput"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [admin]
      summary: Delete a room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /rooms/{id}/images:
    get:
      operationId: listRoomImages
      tags: [roomImages]
      summary: List images for a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomImage"
    post:
      operationId: addRoomImage
      tags: [roomImages]
      summary: Add an image to a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRoomImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomImage"

  /rooms/{id}/images/{imageId}:
    delete:
      operationId: deleteRoomImage
      tags: [roomImages]
      summary: Delete a room image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}:
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /dining:
    get:
      operationId: listDining
      tags: [dining]
      summary: List all dining items
      responses:
        "200":
          description: Dining items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiningItem"
    post:
      operationId: createDiningItem
      tags: [admin]
      summary: Create a dining item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiningItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningItem"

  /dining/{id}:
    put:
      operationId: updateDiningItem
      tags: [admin]
      summary: Update a dining item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiningItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningItem"
    delete:
      operationId: deleteDiningItem
      tags: [admin]
      summary: Delete a dining item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get hotel settings (social media, contact info)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelSettings"
    put:
      operationId: updateSettings
      tags: [admin]
      summary: Update hotel settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelSettings"

  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI hotel agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]

    RoomImage:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        imageUrl:
          type: string
        altText:
          type: string
          nullable: true
        displayOrder:
          type: integer
        createdAt:
          type: string
      required: [id, roomId, imageUrl, displayOrder, createdAt]

    AddRoomImageInput:
      type: object
      properties:
        imageUrl:
          type: string
        altText:
          type: string
          nullable: true
        displayOrder:
          type: integer
      required: [imageUrl]

    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [standard, deluxe, suite, penthouse]
        pricePerNight:
          type: number
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        available:
          type: boolean
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/RoomImage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, pricePerNight, capacity, amenities, available, description, images, createdAt, updatedAt]

    CreateRoomInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [standard, deluxe, suite, penthouse]
        pricePerNight:
          type: number
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        available:
          type: boolean
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [name, type, pricePerNight, capacity, description]

    Booking:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        totalPrice:
          type: number
        status:
          type: string
          enum: [confirmed, cancelled, pending]
        createdAt:
          type: string
        updatedAt:
          type: string
        room:
          $ref: "#/components/schemas/Room"
          nullable: true
      required: [id, roomId, guestName, guestEmail, checkIn, checkOut, totalPrice, status, createdAt, updatedAt]

    CreateBookingInput:
      type: object
      properties:
        roomId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
      required: [roomId, guestName, guestEmail, checkIn, checkOut]

    DiningItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [breakfast, lunch, dinner, drinks, dessert]
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        available:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, description, category, price, available, displayOrder, createdAt, updatedAt]

    CreateDiningItemInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [breakfast, lunch, dinner, drinks, dessert]
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        available:
          type: boolean
        displayOrder:
          type: integer
      required: [name, description, category, price]

    HotelSettings:
      type: object
      properties:
        facebookUrl:
          type: string
        twitterUrl:
          type: string
        instagramUrl:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        hotelName:
          type: string
      required: []

    AiChatInput:
      type: object
      properties:
        message:
          type: string
      required: [message]

    AiChatResponse:
      type: object
      properties:
        response:
          type: string
        tool:
          type: string
          nullable: true
        data:
          nullable: true
      required: [response]

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, size, contentType]

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required: [uploadURL, objectPath]