# lib/api-zod/src/generated/types/addRoomImageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking System API
 * OpenAPI spec version: 0.1.0
 */

export interface AddRoomImageInput {
  imageUrl: string;
  altText?: string | null;
  displayOrder?: number;
}