# lib/api-zod/src/generated/types/createDiningItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDiningItemInputCategory } from "./createDiningItemInputCategory";

export interface CreateDiningItemInput {
  name: string;
  description: string;
  category: CreateDiningItemInputCategory;
  price: number;
  imageUrl?: string | null;
  available?: boolean;
  displayOrder?: number;
}