# lib/api-zod/src/generated/types/createRoomInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRoomInputType } from "./createRoomInputType";

export interface CreateRoomInput {
  name: string;
  type: CreateRoomInputType;
  pricePerNight: number;
  capacity: number;
  amenities?: string[];
  available?: boolean;
  description: string;
  imageUrl?: string | null;
}